anidado barril de componentes de producto y se añadio metadatos a catálogo
