+++ frontend/src/components/index.ts
@@ -2,7 +2,7 @@
 export { default as Wrapper } from "./Layouts/Wrapper/Wrapper";
 export { default as Navbar } from "./Layouts/Navbar/Navbar";
 export { default as Footer } from "./Layouts/Footer/Footer";
-export { default as  WhatsAppLink } from "./Layouts/WhatsAppLink/WhatsAppLink";
+export { default as WhatsAppLink } from "./Layouts/WhatsAppLink/WhatsAppLink";
 
 //Exportaciones Card
 export { default as HomeCard } from "./Cards/HomeCard/HomeCard";
@@ -12,6 +12,17 @@ export { default as PrizeCard } from "./Cards/AwardsCard/PrizeCard";
 export { default as AwardsCards } from "./Cards/AwardsCard/AwardsCards";
 export { default as AboutCard } from "./Cards/AboutCard/AboutCard";
 export { default as AboutSection } from "./Cards/AboutCard/AboutSection";
+export { default as ProductCard } from "./Cards/ProductCard/ProductCard";
+
 
 //Exportaciones Accordion
 export { default as Accordion } from "./Accordion/Accordion";
+
+
+//Exportación FilterBar
+export { default as FilterBar } from "./FilterBar/FilterBar";
+
+//Componentes de productos
+
+export { default as CatalogComponent } from "./CatalogComponents/Catalog"
+export { default as DetailComponent } from "./CatalogComponents/Detail"
